kinematic: Add second collection that tracks dynamic colliders.
