--- VisualPinball.Unity/VisualPinball.Unity/Physics/NativeCollider.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 264d4febde7b4f19955044bba6cbd6d6
-timeCreated: 1698263847
