+++ VisualPinball.Unity/VisualPinball.Unity/Physics/NativeColliderIds.cs
@@ -1,0 +1,7 @@
+﻿namespace VisualPinball.Unity
+{
+	public enum NativeColliderIds
+	{
+		
+	}
+}
