+++ VisualPinball.Unity/VisualPinball.Unity/VPT/Bumper/BumperApi.cs
@@ -72,7 +72,8 @@ namespace VisualPinball.Unity
 		protected override bool FireHitEvents => ColliderComponent.HitEvent;
 		protected override float HitThreshold => ColliderComponent.Threshold;
 
-		protected override void CreateColliders(ref ColliderReference colliders, float margin)
+		protected override void CreateColliders(ref ColliderReference colliders,
+			ref ColliderReference kinematicColliders, float margin)
 		{
 			var height = MainComponent.PositionZ;
 			colliders.Add(new CircleCollider(MainComponent.Position, MainComponent.Radius, height,
