+++ VisualPinball.Unity/VisualPinball.Unity/VPT/CollidableApi.cs
@@ -41,14 +41,15 @@ namespace VisualPinball.Unity
 		protected virtual bool FireHitEvents => false;
 		protected virtual float HitThreshold => 0;
 
-		protected abstract void CreateColliders(ref ColliderReference colliders, float margin);
+		protected abstract void CreateColliders(ref ColliderReference colliders, ref ColliderReference kinematicColliders, float margin);
 
-		void IApiColliderGenerator.CreateColliders(ref ColliderReference colliders, float margin)
+		void IApiColliderGenerator.CreateColliders(ref ColliderReference colliders, ref ColliderReference kinematicColliders,
+			float margin)
 		{
 			if (!ColliderComponent) {
 				return;
 			}
-			CreateColliders(ref colliders, margin);
+			CreateColliders(ref colliders, ref kinematicColliders, margin);
 		}
 
 		ColliderInfo IApiColliderGenerator.GetColliderInfo() => GetColliderInfo(MainComponent.ItemType);
