+++ VisualPinball.Unity/VisualPinball.Unity/VPT/HitTarget/HitTargetApi.cs
@@ -59,7 +59,8 @@ namespace VisualPinball.Unity
 		protected override bool FireHitEvents => true;
 		protected override float HitThreshold => ColliderComponent.Threshold;
 
-		protected override void CreateColliders(ref ColliderReference colliders, float margin)
+		protected override void CreateColliders(ref ColliderReference colliders,
+			ref ColliderReference kinematicColliders, float margin)
 		{
 			var colliderGenerator = new HitTargetColliderGenerator(this, MainComponent, MainComponent);
 			colliderGenerator.GenerateColliders(ref colliders);
