+++ VisualPinball.Unity/VisualPinball.Unity/VPT/ICollidableComponent.cs
@@ -18,7 +18,7 @@ namespace VisualPinball.Unity
 {
 	public interface ICollidableComponent
 	{
-		internal void GetColliders(Player player, ref ColliderReference colliders, float margin);
+		internal void GetColliders(Player player, ref ColliderReference colliders, ref ColliderReference kinematicColliders, float margin);
 		internal int ItemId { get; }
 		internal bool IsCollidable { get; }
 	}
