+++ VisualPinball.Unity/VisualPinball.Unity/VPT/Kicker/KickerApi.cs
@@ -224,7 +224,8 @@ namespace VisualPinball.Unity
 
 		#region Collider Generation
 
-		protected override void CreateColliders(ref ColliderReference colliders, float margin)
+		protected override void CreateColliders(ref ColliderReference colliders,
+			ref ColliderReference kinematicColliders, float margin)
 		{
 			var height = MainComponent.PositionZ;
