+++ VisualPinball.Unity/VisualPinball.Unity/VPT/Playfield/PlayfieldApi.cs
@@ -29,7 +29,8 @@ namespace VisualPinball.Unity
 
 		#region Collider Generation
 
-		protected override void CreateColliders(ref ColliderReference colliders, float margin)
+		protected override void CreateColliders(ref ColliderReference colliders,
+			ref ColliderReference kinematicColliders, float margin)
 		{
 			var info = ((IApiColliderGenerator)this).GetColliderInfo();
