+++ VisualPinball.Unity/VisualPinball.Unity/VPT/Primitive/PrimitiveApi.cs
@@ -43,10 +43,11 @@ namespace VisualPinball.Unity
 		protected override bool FireHitEvents => ColliderComponent.HitEvent;
 		protected override float HitThreshold => ColliderComponent.Threshold;
 
-		protected override void CreateColliders(ref ColliderReference colliders, float margin)
+		protected override void CreateColliders(ref ColliderReference colliders, ref ColliderReference kinematicColliders, float margin)
 		{
 			var colliderGenerator = new PrimitiveColliderGenerator(this, MainComponent, MainComponent);
-			colliderGenerator.GenerateColliders(ColliderComponent.CollisionReductionFactor, ref colliders);
+			var colls = ColliderComponent.IsKinematic ? kinematicColliders : colliders;
+			colliderGenerator.GenerateColliders(ColliderComponent.CollisionReductionFactor, ref colls);
 		}
 
 		#endregion
