+++ VisualPinball.Unity/VisualPinball.Unity/VPT/Spinner/SpinnerApi.cs
@@ -82,7 +82,8 @@ namespace VisualPinball.Unity
 
 		protected override bool FireHitEvents => true;
 
-		protected override void CreateColliders(ref ColliderReference colliders, float margin)
+		protected override void CreateColliders(ref ColliderReference colliders,
+			ref ColliderReference kinematicColliders, float margin)
 		{
 				var colliderGenerator = new SpinnerColliderGenerator(this, MainComponent);
 				colliderGenerator.GenerateColliders(MainComponent.HeightOnPlayfield, ref colliders);
